refactor : RacingGameResult view 패키지로 이동
