+++ src/test/java/racingcar/view/RacingGameResultTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,6 +6,8 @@ import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.RacingCar;
+import racingcar.view.RacingGameResult;
 
 public class RacingGameResultTest {
